backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.3
# PostgreSQL dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1 